analysis-service-python/app/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
